[ADD] cleanup database on deletes

+++ core/database/delete.py
@@ -5,6 +5,8 @@ from core.database.get import (
     get_follower,
     get_inspector,
     get_inspected_user,
+    get_all_inspectors,
+    get_all_inspected_users,
 )
 from core.database.utils import Cursor
 from core.datatypes import Inspector, UnderInspect
@@ -57,12 +59,17 @@ def delete_inspector(inspector: [str, Inspector]) -> Inspector:
 
     inspector_object = get_inspector(inspector)
 
-    with Cursor() as cursor:
-        cursor.execute(
-            "DELETE FROM inspectors WHERE username=?",
-            (inspector,),
-        )
-        cursor.connection.commit()
+    if inspector_object:
+        with Cursor() as cursor:
+            cursor.execute(
+                "DELETE FROM inspectors WHERE username=?",
+                (inspector,),
+            )
+            cursor.execute(
+                "UPDATE under_inspect SET inspector=NULL WHERE inspector=?",
+                (inspector_object.id,),
+            )
+            cursor.connection.commit()
 
     inspector_object.id = None
     return inspector_object
@@ -74,12 +81,32 @@ def delete_inspected_user(inspected_user: [str, UnderInspect]) -> UnderInspect:
 
     inspected_user_object = get_inspected_user(inspected_user)
 
-    with Cursor() as cursor:
-        cursor.execute(
-            "DELETE FROM under_inspect WHERE username=?",
-            (inspected_user,),
-        )
-        cursor.connection.commit()
+    if inspected_user_object:
+        with Cursor() as cursor:
+            cursor.execute(
+                "DELETE FROM under_inspect WHERE username=?",
+                (inspected_user,),
+            )
+            cursor.execute(
+                "DELETE FROM followers WHERE inspected_user=?",
+                (inspected_user_object.id,),
+            )
+            cursor.execute(
+                "DELETE FROM followings WHERE inspected_user=?",
+                (inspected_user_object.id,),
+            )
+            cursor.connection.commit()
 
     inspected_user_object.id = None
     return inspected_user_object
+
+
+def prune_database():
+    inspectors = get_all_inspectors()
+    inspectors_ids = list(map(lambda i: i.id, inspectors))
+
+    inspected_users = get_all_inspected_users()
+
+    for inspected_user in inspected_users:
+        if inspected_user.inspector not in inspectors_ids:
+            delete_inspected_user(inspected_user)
